+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- Thread safety: Change 'candidates' from init/property to `Mapper.apply` arg
 - Remove `store`/`restore` serialize argument. Default path is now None for `Translator.store`.
 
 ## [0.11.1] - 2022-07-02